Static Instance Class is extended with Private and Private Persistent abstract classes
I needed some Static Instances that are not accessible from outside but can check if there is another instance of them. Classes Such as Scene Controller can only be accessed with events but there needs to be only one instance and that instance needs to be persistent. So I created Private Singleton and Private Persistent Singleton abstract classes so that classes that needs this type of functionality can inherit them.