Unused Global and Local Post Process assets are removed, Level 2 and Level 3 Navmesh and Reflection probes are removed
I was trying to solve the issue of UI getting too shiny with global bloom and I tried to solve it with using multiple local Post Processing volumes. None of them properly worked since my UI is also in world space. So I managed to solve it with changing the bloom's threshold and increasing HDR intensities of non-UI Game Objects.