Rock Compute Shader is created
It will rotate the mesh on  a regular interval and lerp its position to next mesh position so that all of the rocks will slowly move. Actually this creates a long line of rocks after some time but cinematic scene ends after 2 minutes so that's not a problem for now.